- Added script to compile c-files into bitcode. - Incorporated script into maven build process

+++ benchmarks/src/main/java/de/consol/dus/graal/numbercrunching/fibonacci/CFibonacciRecursive.java
@@ -21,9 +21,8 @@ public class CFibonacciRecursive extends PolyglotFibonacci {
     }
   }
 
-
   @Override
   public long calculateNthFibonacci(int n) {
-    return function.apply(n).longValue();
+    return function.apply(n);
   }
 }
